Move maxTriesTransfer() from uploader to main Facebook service class
